fix(ssl untrusted cert): throw more helpful error when SSL cert is untrusted
currently, if the server returns an error that the SSL certificate is untrusted, we get an archaic
NSError and a response of "nil". This causes the core to return a "RestError.noResponse" to the
user, which masks the real issue. This adds better checking and handling for SSL cert errors, and
returns a user friendly error.